Extracted side effect from file validity check
I'd rather have the error not be reported by checking function,
but instead where it is called.
It's still not quite optimal, as it now does not provide any different
message if we run into the timeout of `ffprobe` or it actually marking
the given file as invalid, but... good enough.